+++ app/composables/render/idb.ts
@@ -2,6 +2,7 @@ import type { VatSpyAPIData } from '~/types/data/vatspy';
 import type { Table } from 'dexie';
 import Dexie from 'dexie';
 import type {
+    RadarDataAirline,
     RadarDataAirlinesAllList,
     SimAwareAPIData,
     SimAwareDataFeature,
@@ -33,7 +34,9 @@ export type IDBNavigraphProcedures = {
 class VatsimRadarDB extends Dexie {
     data!: Table<VatSpyAPIData | SimAwareAPIData | VatglassesAPIData | IDBAirlinesData, string>;
 
-    keyVal!: Table<SimAwareDataFeature[] | string, `simaware-${ string }` | 'simawareVersion'>;
+    simaware!: Table<SimAwareDataFeature[] | string, string>;
+
+    airlines!: Table<RadarDataAirline | string, string>;
 
     navigraphAirports!: Table<IDBNavigraphProcedures, string>;
 
@@ -49,10 +52,11 @@ export async function initClientDB() {
     indexedDB.deleteDatabase('vatsim-radar');
     const db = new VatsimRadarDB('vatsim-radar-db');
 
-    db.version(2)
+    db.version(3)
         .stores({
             data: '',
-            keyVal: '',
+            simaware: '',
+            airlines: '',
             navigraphAirports: '',
             navigraphData: '',
             navigraphDB: '',