+++ app/utils/map/entities.ts
@@ -7,6 +7,7 @@ import type VectorSource from 'ol/source/Vector';
 import type { SimAwareProperties } from '~/utils/server/storage';
 import type { VatsimShortenedController, VatsimShortenedPrefile } from '~/types/data/vatsim';
 import type { MapAircraftKeys, MapAircraftList } from '~/types/map';
+import type { Coordinate } from 'ol/coordinate.js';
 
 export const globalMapEntities = {
     airports: null as VectorSource | null,
@@ -22,6 +23,7 @@ export interface FeatureAirportProperties {
     lon: number;
     atc: VatsimShortenedController[];
     atcLength: number;
+    isPseudo: boolean;
     aircraftList: MapAircraftList;
 
     id: `airport-${ string }`;
@@ -107,6 +109,33 @@ export interface FeatureAirportNavigraphProperties {
     id: `airport-${ string }-${ 'gate' | 'runway' | 'layer' }-${ string }`;
 }
 
+export interface FeatureAirportSectorDefaultProperties {
+    id: `sector-${ string }`;
+    type: 'sector';
+    sectorType: 'empty' | 'fir' | 'uir';
+    selected?: boolean;
+    booked: boolean;
+    duplicated: boolean;
+    icao: string;
+    uir?: string;
+    label: Coordinate;
+    atc: VatsimShortenedController[];
+}
+
+export interface FeatureAirportSectorVGProperties {
+    type: 'sector-vatglasses';
+    sectorType: 'vatglasses';
+    id?: never;
+    vgSectorId: string;
+    selected?: boolean;
+    min: number;
+    max: number;
+    countryGroupId: string;
+    positionId: string;
+    atc: VatsimShortenedController[];
+    colour: string;
+}
+
 export type FeatureAirport = Feature<Point, FeatureAirportProperties>;
 export type FeatureAirportApproach = Feature<Polygon, FeatureAirportApproachProperties>;
 export type FeatureAirportApproachLabel = Feature<Point, FeatureAirportApproachTextProperties>;
@@ -114,6 +143,8 @@ export type FeatureAirportGate = Feature<Point, FeatureAirportGateProperties>;
 export type FeatureAirportAtc = Feature<Point, FeatureAirportAirportAtcProperties>;
 export type FeatureAirportCounter = Feature<Point, FeatureAirportAirportCounterProperties>;
 export type FeatureAirportNavigraph = Feature<Point, FeatureAirportNavigraphProperties>;
+export type FeatureSector = Feature<Polygon | MultiPolygon, FeatureAirportSectorDefaultProperties>;
+export type FeatureSectorVG = Feature<Polygon | MultiPolygon, FeatureAirportSectorVGProperties>;
 
 export type FeatureSimAware = Feature<Polygon | MultiPolygon, SimAwareProperties & { type: 'simaware' }>;
 
@@ -125,6 +156,8 @@ export type MapFeatures =
     | FeatureAirportAtc
     | FeatureAirportCounter
     | FeatureAirportNavigraph
+    | FeatureSector
+    | FeatureSectorVG
     | FeatureSimAware;
 
 export type MapFeaturesType = ReturnType<MapFeatures['getProperties']>['type'];
@@ -155,7 +188,7 @@ export function getMapFeature<T extends MapFeaturesType>(
     source: VectorSource,
     id: MapFeatureProperties<T>['id'],
 ): FeatureForType<T> | null {
-    return source.getFeatureById(id) as FeatureForType<T> | null;
+    return source.getFeatureById(id!) as FeatureForType<T> | null;
 }
 
 export function isMapFeature<T extends MapFeaturesType>(type: T, properties: Record<string, any>): properties is MapFeatureProperties<T> {