+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+# [2.0.0-alpha.3]
+
+## Alpha changes
+
+- Significantly improved memory usage for airports style
+- Controllers in airport overlay are now sorted by position
+- Fixed overlays enabling on mobile even when map is moving
+- Fixed airports buttons not clickable on mobile
+- Updated atc icons
+- Improved SimAware data update stability
+- FIR/VG list v2 (render rework), no multiselect for now
+- Fixed controller duplication issues
+- "Max counters to show" and "Hide airports on zoom" now also affect TRACONs labels
+
+## Improvements and changes
+
+- Implement on-hover delay
+- Added an ID for VATGlasses sectors
+
+## Performance Improvements
+
+- Significantly reduced memory usage for airlines icao codes
+
 # [2.0.0-alpha.2]
 
 ## Improvements and changes